threshold_scan steps by two and has a longer wait after setting the DACs to allow for the rates to settle and eliminate reading the same singles rates twice.
discone_global_trigger_rate_scan_with_alighed_thresholds adds an array
that shifts the DACs from the printed value as to make their noise
curves more correlated during the scan. It also prints out the L1 and L2
rates and has a longer sleep to print out the 1Hz scalars
correctly. The range it scans is much smaller as well to include the
most interesting parts of the discone channels, to eliminate edge
effects, and to save time.

discone_global_threshold_scan adds a loop to build up statistics for
the 1Hz L1 and L2 scalar rates. It scans an even smaller range.

Makefile was altered to include these two additional scripts